requirements.txt
fastapi==0.115.2
uvicorn==0.31.0
sqlalchemy==2.0.34
psycopg[binary]==3.2.3
pydantic==2.9.2
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
httpx==0.27.2
python-multipart==0.0.9
argon2-cffi==23.1.0